refactor(YouTube): Extract flyout video ID logic into FlyoutUtils
