+++ extensions/music/src/main/java/app/morphe/extension/music/patches/DownloadsPatch.java
@@ -14,7 +14,6 @@ import androidx.annotation.Nullable;
 
 import com.facebook.litho.ComponentHost;
 
-import java.lang.reflect.Method;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
