+++ patches/src/main/kotlin/app/morphe/patches/youtube/layout/sharesheet/OpenSystemShareSheetPatch.kt
@@ -11,7 +11,6 @@ import app.morphe.patcher.extensions.InstructionExtensions.addInstruction
 import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.patches.shared.misc.litho.filter.addLithoFilter
 import app.morphe.patches.shared.misc.settings.preference.SwitchPreference
-import app.morphe.patches.youtube.layout.flyout.addToQueuePatch
 import app.morphe.patches.youtube.misc.extension.sharedExtensionPatch
 import app.morphe.patches.youtube.misc.litho.filter.lithoFilterPatch
 import app.morphe.patches.youtube.misc.recyclerviewtree.addRecyclerViewTreeHook
@@ -37,8 +36,7 @@ internal fun openSystemShareSheetPatch(
         sharedExtensionPatch,
         settingsPatch,
         lithoFilterPatch,
-        recyclerViewTreeHookPatch,
-        addToQueuePatch
+        recyclerViewTreeHookPatch
     )
 
     compatibleWith(COMPATIBILITY_YOUTUBE)
